+++ services/backend/src/app.module.ts
@@ -1,12 +1,19 @@
+import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { EventsController } from './events/events.controller.js';
 import { EventsService } from './events/events.service.js';
-import { Module } from '@nestjs/common';
 import { AppController } from './app.controller.js';
 import { AppService } from './app.service.js';
+import { AuthModule } from './auth/auth.module.js';
+import { AuthController } from './auth/auth.controller.js';
+import { FirebaseAuthenticationMiddleware } from './auth/authentication/firebase-authentication.middleware.js';
 
 @Module({
-  imports: [],
+  imports: [AuthModule],
   controllers: [AppController, EventsController],
   providers: [AppService, EventsService],
 })
-export class AppModule {}
+export class AppModule implements NestModule {
+  configure(consumer: MiddlewareConsumer): void {
+    consumer.apply(FirebaseAuthenticationMiddleware).forRoutes(AuthController);
+  }
+}